rename CustomEdge.tsx to custom-edge.tsx
